Update to PostCSS 8
