--- .babelrc
@@ -1,12 +1,0 @@
-{
-  "presets": [
-    ["env", {
-      "targets": {
-        "node": 4
-      }
-    }]
-  ],
-  "plugins": [
-    "add-module-exports"
-  ]
-}
