+++ .eslintrc
@@ -14,7 +14,6 @@
     "vars-on-top": 0
   },
   "env": {
-    "jest": true,
     "node": true
   }
 }
