--- jsconfig.json
@@ -1,5 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "ES6"
-  }
-}
